Add new implementation of auto-scrolling
This logic is better-suited to the `app` module as it contains
functionality which may be of use to the entire application (and not
only the TranscriptView).

The previous implementation used a boolean flag to track whether a
scroll animation had been programatically triggered. This implementation
tracks the expected position of the scrollbar over time, accounting for
possible error due to throttling and browser scheduling. Under this
approach, the system can recognize user-initiated scroll events that
interrupt system-initiated scroll animations.